threat-model: introduce initial document
This is a starting point, and could use a bit of help and
review from the community. This focuses on the devices / interfaces
with the virtual machine, and ultimately to the container itself.

We then discuss how these devices/interfaces vary by VMM/hypervisor.

The threat model drawing is created via gdocs, located here:
https://docs.google.com/drawings/d/1_lsgchTW-qUwdl_FcivA2fsMu2Q9G5eJM0khF0NE5tE/edit?usp=sharing

Fixes: #572

Signed-off-by: Eric Ernst <eric.ernst@intel.com>